fixing

+++ src/Product_page/review.php
@@ -37,7 +37,7 @@ while($review = $review_result->fetch_assoc()){
         <?php
         echo $user_id;
         echo $review_user_id;
-        if($user_id == $review_user_id || $user_id == 0){
+        if($user_id == $review_user_id && isset($user_id) || $user_id == 0 && isset($user_id)){
           ?>
           <button class="delete_button" value="Delete" onclick="delete_review('<?php echo $review_id ?>')">Delete</button>
           <?php
